+++ Infra/DbManager.Infra.WebApi/DbManager.Infra.WebApi.csproj
@@ -5,4 +5,8 @@
         <LangVersion>default</LangVersion>
     </PropertyGroup>
 
+    <ItemGroup>
+        <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0-rc4" />
+    </ItemGroup>
+
 </Project>
