+++ Infra/DbManager.Infra.WebApi/Extensions/SwaggerExtensions.cs
