+++ dbmanager.sln
@@ -1,8 +1,6 @@
 
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "dbmanager.API", "dbmanager.API\dbmanager.API.csproj", "{18261C69-B0EB-46CE-A2A1-B33CC5612AA7}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{3C950476-4232-408A-8DC5-78A17987F790}"
 	ProjectSection(SolutionItems) = preProject
 		README.md = README.md
@@ -41,12 +39,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{18261C69-B0EB-46CE-A2A1-B33CC5612AA7}.Default|Any CPU.ActiveCfg = Debug|Any CPU
-		{18261C69-B0EB-46CE-A2A1-B33CC5612AA7}.Default|Any CPU.Build.0 = Debug|Any CPU
-		{18261C69-B0EB-46CE-A2A1-B33CC5612AA7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{18261C69-B0EB-46CE-A2A1-B33CC5612AA7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{18261C69-B0EB-46CE-A2A1-B33CC5612AA7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{18261C69-B0EB-46CE-A2A1-B33CC5612AA7}.Release|Any CPU.Build.0 = Release|Any CPU
 		{761C5CA1-C6D3-4B18-A5BA-0E61DFFBB7A0}.Default|Any CPU.ActiveCfg = Debug|Any CPU
 		{761C5CA1-C6D3-4B18-A5BA-0E61DFFBB7A0}.Default|Any CPU.Build.0 = Debug|Any CPU
 		{761C5CA1-C6D3-4B18-A5BA-0E61DFFBB7A0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
